Move peak_picker module into pyFAI.gui
